ref(entropy): renamed Entropy to index
