+++ src/screens/LoggedOut/entropy/index.tsx
