+++ src/screens/LoggedOut/index.tsx
@@ -4,7 +4,7 @@ import { createStackNavigator } from '@react-navigation/stack'
 import { ScreenNames } from '~/types/screens'
 
 import Walkthrough from './Walkthrough'
-import { Entropy } from './entropy/Entropy'
+import { Entropy } from './entropy'
 import Recovery from './Recovery'
 import SeedPhrase from './SeedPhrase'
 import SeedPhraseRepeat from './SeedPhraseRepeat'
